pyproject.toml
[tool.poetry]
name = "groovis"
version = "0.1.0"
description = ""
authors = ["jihoahn9303 <rediate03@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
Pillow = "^9.1.1"
torch = "^1.12.0"
einops = "^0.4.1"
albumentations = "^1.2.1"
matplotlib = "^3.5.2"
torchvision = "^0.13.1"
wandb = "^0.13.3"
python-dotenv = "^0.21.0"
omegaconf = "^2.2.3"
torchmetrics = "^0.9.3"
loguru = "^0.6.0"
datasets = "^2.5.1"
torch-optimizer = "^0.3.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"
pytest = "^7.1.2"
ImageHash = "^4.2.1"
pytest-watch = "^4.2.0"
pytest-testmon = "^1.3.4"
pre-commit = "^2.20.0"
mypy = "^0.982"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
